+++ tests/test__internal/test_init.py
@@ -97,25 +97,10 @@ async def test_gather_checks():
     )
 
 
-@pytest.mark.parametrize(
-    ("content", "prefix", "expected_result"),
-    [
-        ("no go sir", ("no", "home", "blow"), "no"),
-        ("hime", ("hi", "hime", "boomer"), "hime"),
-        ("boomer", ("boo", "boomer", "no u"), "boomer"),
-        ("ok boomer", ("no", "nani"), None),
-        ("", ("nannnnni",), None),
-        ("ok ok ok", (), None),
-    ],
-)
-def test_match_prefix_names(content: str, prefix: str, expected_result: typing.Optional[str]):
-    assert _internal.match_prefix_names(content, prefix) == expected_result
-
-
 class TestCastedView:
     def test___getitem___for_non_existant_entry(self):
         mock_cast = mock.Mock()
-        view = _internal.CastedView[typing.Any, typing.Any, typing.Any]({}, mock_cast)
+        view = _internal.CastedView[typing.Any, typing.Any]({}, mock_cast)
 
         with pytest.raises(KeyError):
             view["foo"]
@@ -125,7 +110,7 @@ class TestCastedView:
     def test___getitem___for_buffered_entry(self):
         mock_cast = mock.Mock()
         mock_value = mock.MagicMock()
-        view = _internal.CastedView[typing.Any, str, typing.Any]({"a": "b"}, mock_cast)
+        view = _internal.CastedView[typing.Any, str]({"a": "b"}, mock_cast)
         view._buffer["a"] = mock_value
 
         result = view["a"]
@@ -136,7 +121,7 @@ class TestCastedView:
     def test___getitem___for_not_buffered_entry(self):
         mock_cast = mock.Mock()
         mock_value = mock.MagicMock()
-        view = _internal.CastedView[typing.Any, typing.Any, typing.Any]({"a": mock_value}, mock_cast)
+        view = _internal.CastedView[typing.Any, typing.Any]({"a": mock_value}, mock_cast)
 
         result = view["a"]
 
@@ -147,13 +132,13 @@ class TestCastedView:
     def test___iter__(self):
         mock_iter = mock.Mock(return_value=iter((1, 2, 3)))
         mock_dict = mock.Mock(__iter__=mock_iter)
-        view = _internal.CastedView[int, int, typing.Any](mock_dict, mock.Mock())
+        view = _internal.CastedView[int, int](mock_dict, mock.Mock())
 
         assert iter(view) is mock_iter.return_value
 
     def test___len___(self):
         mock_dict = mock.Mock(__len__=mock.Mock(return_value=43123))
-        view = _internal.CastedView[int, int, typing.Any](mock_dict, mock.Mock())
+        view = _internal.CastedView[int, int](mock_dict, mock.Mock())
 
         assert len(view) == 43123